+++ Assets/Animations/Timeline/Conversation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5e566826e6a90204eacdbcdab5b2d3ed
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
