+++ Assets/Animations/Timeline/MiniGame.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 14e835f21fa6f024e9f23368bf24d807
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
