+++ Assets/Animations/Timeline/MiniGame/ResultTimeline.playable.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aaceeb4d0ea4f8d4086a41348bfb9a50
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
