+++ Assets/Scenes/MiniGame/MiniGameTest.unity.meta
@@ -1,6 +1,5 @@
 fileFormatVersion: 2
-guid: 599bcd2051e6aa449b8e0dcc8d9903e9
-folderAsset: yes
+guid: 6d0ac4eeb7fd9b44790a82fc01fcaad0
 DefaultImporter:
   externalObjects: {}
   userData:
