--- Assets/Scripts/MiniGame/ItemToMouse.cs
@@ -1,32 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class ItemToMouse : MonoBehaviour
-{
-    [SerializeField]
-    private GameObject m_object = null;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        Vector3 touchScreenPosition = Input.mousePosition;
-
-        touchScreenPosition.x = Mathf.Clamp(touchScreenPosition.x, 0.0f, Screen.width);
-        touchScreenPosition.y = Mathf.Clamp(touchScreenPosition.y, 0.0f, Screen.height);
-
-        // 10.0fに深い意味は無い。画面に表示したいので適当な値を入れてカメラから離そうとしているだけ.
-        touchScreenPosition.z = 10.0f;
-
-        Camera gameCamera = Camera.main;
-        Vector3 touchWorldPosition = gameCamera.ScreenToWorldPoint(touchScreenPosition);
-
-        m_object.transform.position = touchWorldPosition;
-    }
-}
