+++ Assets/Scripts/MiniGame/MiniGameManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2485aa581b275e249bdaa37e76f36ac5
+guid: 4e5c4e8b172b14444bdc62c5cddfcc82
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
