+++ Assets/Scripts/MiniGame/MixManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6f9ef14f36657a24899fd152b2ffa3eb
+guid: 2485aa581b275e249bdaa37e76f36ac5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
