+++ Assets/Scripts/MiniGame/Mix_Wall.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class WallController : MonoBehaviour
+public class Mix_Wall : MonoBehaviour
 {
     private int ClickNum;
     public byte ColorNum;
