+++ Assets/Scripts/MiniGame/PaintManager.cs.meta
